ir: drop dead LoopOp helpers + unused Accumulator.dtype
Further simplification pass following the Axis.kind removal:

- Drop LoopOp.free_axes() and LoopOp.reduce_axes() methods. Zero
  external callers (only free_axes was used internally by
  infer_output_shape, itself now gone). Call sites that need free
  axes inline the filter via reduce_axis_names.

- Drop LoopOp.infer_output_shape() method. No LoopOp callers outside
  loop_ir.py itself (decomposition's infer_output_shape is on ReshapeOp,
  a different class). Inline the one fallback in _infer_write_shape.

- Drop Accumulator.dtype field. Stored but never read — simplify.py's
  rebuild path was the only place that touched it, as a pass-through.

- Inline _merge_accumulators in _merge_core.py. 12-line helper with one
  call site; the two-line inlined version is clearer in context.

Net: -30 lines. All 852 tests pass; make lint clean.

+++ deplodock/compiler/ir/simplify.py
@@ -391,15 +391,7 @@ def simplify_loop_op(op: LoopOp) -> LoopOp:
     # top-level inputs/accumulators we rebuild with an empty Context first
     # and rely on the Loop walker to push axis ranges.
     new_inputs = tuple(Port(_simplify_expr_tuple(p.index, ctx)) for p in op.inputs)
-    new_accumulators = tuple(
-        Accumulator(
-            name=acc.name,
-            combine=acc.combine,
-            init=simplify_expr(acc.init, ctx),
-            dtype=acc.dtype,
-        )
-        for acc in op.accumulators
-    )
+    new_accumulators = tuple(Accumulator(name=acc.name, combine=acc.combine, init=simplify_expr(acc.init, ctx)) for acc in op.accumulators)
     # Body — push axis ranges as we descend into Loop blocks.
     new_body = tuple(_simplify_loop_stmt(s, ctx) for s in op.body)
     # Now walk inputs again with a Context that has every axis range, so
